feat: move schema to lib
